refactor(core): Extract TriggerExecutionContextFactory from ActiveWorkflowManager
Moves getExecuteTriggerFunctions, getExecutePollFunctions, executeErrorWorkflow
wrapper, and loadPublishedWorkflowData into a new TriggerExecutionContextFactory
service. ActiveWorkflowManager keeps thin delegates and injects the legacy
onTriggerFailure callback for path-specific failure handling.

https://linear.app/n8n/issue/CAT-3395
