+++ packages/cli/src/active-workflow-manager.ts
@@ -16,20 +16,13 @@ import {
 	ActiveWorkflowTriggers,
 	ErrorReporter,
 	InstanceSettings,
-	PollContext,
-	StorageConfig,
-	TriggerContext,
 	type IGetExecutePollFunctions,
 	type IGetExecuteTriggerFunctions,
 } from 'n8n-core';
 import type {
 	ExecutionError,
 	IConnections,
-	IDeferredPromise,
-	IExecuteResponsePromiseData,
 	INode,
-	INodeExecutionData,
-	IRun,
 	IWorkflowBase,
 	IWorkflowExecuteAdditionalData,
 	WorkflowActivateMode,
@@ -44,18 +37,12 @@ import {
 	UnexpectedError,
 	IsolateError,
 	ensureError,
-	createRunExecutionData,
 	validateWorkflowHasTriggerLikeNode,
 } from 'n8n-workflow';
 import { strict } from 'node:assert';
 
 import { ActivationErrorsService } from '@/activation-errors.service';
-import { DuplicateExecutionError } from '@/errors/duplicate-execution.error';
 import { MessageEventBus } from '@/eventbus/message-event-bus/message-event-bus';
-import { ActiveExecutions } from '@/active-executions';
-import { EventService } from '@/events/event.service';
-import { executeErrorWorkflow } from '@/execution-lifecycle/execute-error-workflow';
-import { ExecutionService } from '@/executions/execution.service';
 import { ExternalHooks } from '@/external-hooks';
 import { NodeTypes } from '@/node-types';
 import { Push } from '@/push';
@@ -65,9 +52,8 @@ import { ActiveWorkflowsService } from '@/services/active-workflows.service';
 import * as WebhookHelpers from '@/webhooks/webhook-helpers';
 import { WebhookService } from '@/webhooks/webhook.service';
 import * as WorkflowExecuteAdditionalData from '@/workflow-execute-additional-data';
-import { WorkflowExecutionService } from '@/workflows/workflow-execution.service';
-import { WorkflowPublishedDataService } from '@/workflows/workflow-published-data.service';
 import { WorkflowStaticDataService } from '@/workflows/workflow-static-data.service';
+import { TriggerExecutionContextFactory } from '@/workflows/triggers/trigger-execution-context.factory';
 import { getErrorDescription, getErrorNodeId } from '@/workflows/utils';
 import { formatWorkflow } from '@/workflows/workflow.formatter';
 
@@ -86,23 +72,18 @@ export class ActiveWorkflowManager {
 		private readonly logger: Logger,
 		private readonly errorReporter: ErrorReporter,
 		private readonly activeWorkflowTriggers: ActiveWorkflowTriggers,
-		private readonly activeExecutions: ActiveExecutions,
 		private readonly externalHooks: ExternalHooks,
 		private readonly nodeTypes: NodeTypes,
 		private readonly webhookService: WebhookService,
 		private readonly workflowRepository: WorkflowRepository,
 		private readonly activationErrorsService: ActivationErrorsService,
-		private readonly executionService: ExecutionService,
 		private readonly workflowStaticDataService: WorkflowStaticDataService,
 		private readonly activeWorkflowsService: ActiveWorkflowsService,
-		private readonly workflowExecutionService: WorkflowExecutionService,
 		private readonly instanceSettings: InstanceSettings,
 		private readonly publisher: Publisher,
 		private readonly workflowsConfig: WorkflowsConfig,
 		private readonly push: Push,
-		private readonly eventService: EventService,
-		private readonly storageConfig: StorageConfig,
-		private readonly workflowPublishedDataService: WorkflowPublishedDataService,
+		private readonly triggerExecutionContextFactory: TriggerExecutionContextFactory,
 		private readonly eventBus: MessageEventBus,
 	) {
 		this.logger = this.logger.scoped(['workflow-activation']);
@@ -341,58 +322,15 @@ export class ActiveWorkflowManager {
 		additionalData: IWorkflowExecuteAdditionalData,
 		mode: WorkflowExecuteMode,
 		activation: WorkflowActivateMode,
-		// TODO(CAT-3202): this callback lets us switch between reading from
-		// the in-memory workflowData (flag off) and the workflow published data
-		// service (flag on). Once the feature flag is removed, we'll call the
-		// service directly and this parameter will go away.
 		resolveWorkflowData: () => Promise<IWorkflowBase>,
 	): IGetExecutePollFunctions {
-		return (workflow: Workflow, node: INode) => {
-			const __emit = (
-				data: INodeExecutionData[][],
-				responsePromise?: IDeferredPromise<IExecuteResponsePromiseData>,
-				donePromise?: IDeferredPromise<IRun | undefined>,
-			) => {
-				this.logger.debug(`Received event to trigger execution for workflow "${workflow.name}"`);
-				void this.workflowStaticDataService.saveStaticData(workflow);
-
-				// TODO(CAT-3202): resolves workflow data via callback so we
-				// can feature-flag between in-memory data and the published data
-				// service. Once the flag is removed, we'll call the service directly.
-				const executePromise = resolveWorkflowData().then(
-					async (freshWorkflowData) =>
-						await this.workflowExecutionService.runWorkflow(
-							freshWorkflowData,
-							node,
-							data,
-							additionalData,
-							mode,
-							responsePromise,
-						),
-				);
-
-				if (donePromise) {
-					void executePromise.then((executionId) => {
-						this.activeExecutions
-							.getPostExecutePromise(executionId)
-							.then(donePromise.resolve)
-							.catch(donePromise.reject);
-					});
-				} else {
-					void executePromise.catch((error: Error) => this.logger.error(error.message, { error }));
-				}
-			};
-
-			const __emitError = (error: ExecutionError) => {
-				void this.executionService
-					.createErrorExecution(error, node, workflowData, workflow, mode)
-					.then(() => {
-						this.executeErrorWorkflow(error, workflowData, mode);
-					});
-			};
-
-			return new PollContext(workflow, node, additionalData, mode, activation, __emit, __emitError);
-		};
+		return this.triggerExecutionContextFactory.getExecutePollFunctions(
+			workflowData,
+			additionalData,
+			mode,
+			activation,
+			resolveWorkflowData,
+		);
 	}
 
 	/**
@@ -404,180 +342,41 @@ export class ActiveWorkflowManager {
 		additionalData: IWorkflowExecuteAdditionalData,
 		mode: WorkflowExecuteMode,
 		activation: WorkflowActivateMode,
-		// TODO(CAT-3202): this callback lets us switch between reading from
-		// the in-memory workflowData (flag off) and the workflow published data
-		// service (flag on). Once the feature flag is removed, we'll call the
-		// service directly and this parameter will go away.
 		resolveWorkflowData: () => Promise<IWorkflowBase>,
 	): IGetExecuteTriggerFunctions {
-		return (workflow: Workflow, node: INode) => {
-			const emit = (
-				data: INodeExecutionData[][],
-				responsePromise?: IDeferredPromise<IExecuteResponsePromiseData>,
-				donePromise?: IDeferredPromise<IRun | undefined>,
-				deduplicationKey?: string,
-			) => {
-				this.logger.debug(`Received trigger for workflow "${workflow.name}"`);
-				void this.workflowStaticDataService.saveStaticData(workflow);
-
-				// TODO(CAT-3202): resolves workflow data via callback so we
-				// can feature-flag between in-memory data and the published data
-				// service. Once the flag is removed, we'll call the service directly.
-				const executePromise = resolveWorkflowData()
-					.then(
-						async (freshWorkflowData) =>
-							await this.workflowExecutionService.runWorkflow(
-								freshWorkflowData,
-								node,
-								data,
-								additionalData,
-								mode,
-								responsePromise,
-								deduplicationKey,
-							),
-					)
-					.catch((error: unknown) => {
-						if (error instanceof DuplicateExecutionError) {
-							const context = {
-								workflowId: workflowData.id,
-								nodeId: node.id,
-								deduplicationKey: error.deduplicationKey,
-							};
-							this.logger.warn('Scheduled execution skipped: duplicate deduplication key', context);
-							this.errorReporter.warn(error, { extra: context, shouldBeLogged: false });
-							return undefined;
-						}
-						throw error;
-					});
-
-				void executePromise.then((executionId) => {
-					// `executionId` is undefined when the catch above swallowed a
-					// duplicate scheduled execution; nothing ran, so nothing to emit.
-					if (executionId === undefined) return;
-					this.eventService.emit('workflow-executed', {
-						workflowId: workflowData.id,
-						workflowName: workflowData.name,
-						executionId,
-						source: 'trigger',
-					});
-				});
-
-				if (donePromise) {
-					void executePromise.then((executionId) => {
-						// Same as above: a duplicate scheduled execution was skipped,
-						// so resolve with undefined and don't wait on a non-existent run.
-						if (executionId === undefined) {
-							donePromise.resolve(undefined);
-							return;
-						}
-						this.activeExecutions
-							.getPostExecutePromise(executionId)
-							.then(donePromise.resolve)
-							.catch(donePromise.reject);
-					});
-				} else {
-					executePromise.catch((error: Error) => this.logger.error(error.message, { error }));
-				}
-			};
-			const emitError = (error: Error): void => {
+		return this.triggerExecutionContextFactory.getExecuteTriggerFunctions(
+			workflowData,
+			additionalData,
+			mode,
+			activation,
+			resolveWorkflowData,
+			(error, node, wd, m, act) => {
 				this.logger.info(
-					`The trigger node "${node.name}" of workflow "${workflowData.name}" failed with the error: "${error.message}". Will try to reactivate.`,
+					`The trigger node "${node.name}" of workflow "${wd.name}" failed with the error: "${error.message}". Will try to reactivate.`,
 					{
 						nodeName: node.name,
-						workflowId: workflowData.id,
-						workflowName: workflowData.name,
+						workflowId: wd.id,
+						workflowName: wd.name,
 					},
 				);
-
-				// Remove the workflow as "active"
-
-				void this.activeWorkflowTriggers.remove(workflowData.id);
-
-				void this.activationErrorsService.register(workflowData.id, error.message);
-
-				// Run Error Workflow if defined
+				void this.activeWorkflowTriggers.remove(wd.id);
+				void this.activationErrorsService.register(wd.id, error.message);
 				const activationError = new WorkflowActivationError(
 					`There was a problem with the trigger node "${node.name}", for that reason did the workflow had to be deactivated`,
 					{ cause: error, node },
 				);
-				this.executeErrorWorkflow(activationError, workflowData, mode);
-
-				this.addQueuedWorkflowActivation(activation, workflowData as WorkflowEntity);
-			};
-
-			const saveFailedExecution = (error: ExecutionError) => {
-				this.logger.info(
-					`The trigger node "${node.name}" of workflow "${workflowData.name}" reported the error: "${error.message}". Saving to failed executions`,
-					{
-						nodeName: node.name,
-						workflowId: workflowData.id,
-						workflowName: workflowData.name,
-					},
-				);
-				void this.executionService
-					.createErrorExecution(error, node, workflowData, workflow, mode)
-					.then(() => {
-						this.executeErrorWorkflow(error, workflowData, mode);
-					});
-			};
-			return new TriggerContext(
-				workflow,
-				node,
-				additionalData,
-				mode,
-				activation,
-				emit,
-				emitError,
-				saveFailedExecution,
-			);
-		};
+				this.executeErrorWorkflow(activationError, wd, m);
+				this.addQueuedWorkflowActivation(act, wd as WorkflowEntity);
+			},
+		);
 	}
 
 	executeErrorWorkflow(
 		error: ExecutionError,
 		workflowData: IWorkflowBase,
 		mode: WorkflowExecuteMode,
-	) {
-		const fullRunData: IRun = {
-			data: createRunExecutionData({
-				resultData: {
-					error,
-					runData: {},
-				},
-			}),
-			finished: false,
-			mode,
-			startedAt: new Date(),
-			stoppedAt: new Date(),
-			status: 'running',
-			storedAt: this.storageConfig.modeTag,
-		};
-
-		executeErrorWorkflow(workflowData, fullRunData, mode);
-	}
-
-	/**
-	 * Load the published workflow nodes/connections from the
-	 * `workflow_published_version` table. The passed-in workflow is used
-	 * for all other fields (staticData, settings, etc.).
-	 *
-	 * TODO: Add error handling / fallback strategy for transient DB failures.
-	 */
-	private async loadPublishedWorkflowData(
-		initialWorkflowData: IWorkflowDb,
-	): Promise<IWorkflowBase> {
-		const publishedData = await this.workflowPublishedDataService.getPublishedWorkflowData(
-			initialWorkflowData.id,
-		);
-
-		if (!publishedData) {
-			throw new UnexpectedError('Published version not found for workflow', {
-				extra: { workflowId: initialWorkflowData.id },
-			});
-		}
-
-		const { nodes, connections } = publishedData.publishedVersion;
-		return { ...initialWorkflowData, nodes, connections };
+	): void {
+		this.triggerExecutionContextFactory.executeErrorWorkflow(error, workflowData, mode);
 	}
 
 	private isActivationInProgress = false;
@@ -829,7 +628,8 @@ export class ActiveWorkflowManager {
 				// removed/disabled triggers in a new workflow version are deactivated before
 				// updating the published version.
 				const resolveWorkflowData = this.workflowsConfig.useWorkflowPublicationService
-					? async () => await this.loadPublishedWorkflowData(dbWorkflow)
+					? async () =>
+							await this.triggerExecutionContextFactory.loadPublishedWorkflowData(dbWorkflow)
 					: async () => dbWorkflow as IWorkflowBase;
 
 				if (shouldAddNonWebhookTriggers) {
@@ -929,7 +729,8 @@ export class ActiveWorkflowManager {
 
 			if (this.shouldAddNonWebhookTriggers()) {
 				const resolveWorkflowData = this.workflowsConfig.useWorkflowPublicationService
-					? async () => await this.loadPublishedWorkflowData(dbWorkflow)
+					? async () =>
+							await this.triggerExecutionContextFactory.loadPublishedWorkflowData(dbWorkflow)
 					: async () => dbWorkflow as IWorkflowBase;
 
 				await this.addNonWebhookTriggers(dbWorkflow, workflow, {
